+++ core/exports/effects.ts
@@ -7,6 +7,6 @@ export {
 } from "../src/effects/effects.ts";
 export { io } from "../src/effects/io.ts";
 export { config, denoConfig } from "../src/effects/config.ts";
-export { importmap } from "../src/effects/impormap.ts";
+export { importmap } from "../src/effects/importmap.ts";
 export { hot } from "../src/effects/hot-update.ts";
 export { Handler } from "../src/effects/handlers.ts";
